requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
bigquery==0.0.20
boto3==1.20.26
botocore==1.23.26
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
dacktool==0.0.7
dbstream==0.1.13
Django==3.2.6
django-cors-headers==3.10.1
djangorestframework==3.13.1
ffmpeg-python==0.2.0
future==0.18.2
google-api-core==2.3.2
google-api-python-client==1.7.11
google-auth==2.3.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.2.1
google-cloud-secret-manager==2.7.2
google-cloud-storage==1.43.0
google-cloud-vision==2.6.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
googleauthentication==0.0.16
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
jmespath==0.10.0
libcst==0.3.23
mypy-extensions==0.4.3
numpy==1.21.5
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
paramiko==2.9.1
proto-plus==1.19.8
protobuf==3.19.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyNaCl==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
sshtunnel==0.1.5
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.7
